Allow users to inject extra `deps` into pip packages
There is currently no way to inject a dependency into a pip package.

For example, there is code in the `hypothesis` package that expects
either `importlib_metadata` or `setuptools` to be installed when running
under Python 3.7. Neither of those packages is listed as a dependency
of `hypothesis`. The package is therefore not entirely functional as-is.
Users have to manually depend on one of these 2 dependencies in order
for `hypothesis` to work.

Code in question:
https://github.com/HypothesisWorks/hypothesis/blob/fda07ec3fba53572b5f729aefd16aceb1e61f638/hypothesis-python/src/hypothesis/entry_points.py

Without this new feature, the user will see the following warning when
importing `hypothesis`.
> Under Python <= 3.7, Hypothesis requires either the importlib_metadata
> or setuptools package in order to load plugins via entrypoints.

Rather than making users deal with it, I'd much rather be able to
add it as an annotation and then have it "just work" for everyone.